Fix error message for folder IAM test
