+++ google/resource_google_folder_iam_binding_test.go
@@ -5,10 +5,12 @@ import (
 	"sort"
 	"testing"
 
+	"github.com/hashicorp/errwrap"
 	"github.com/hashicorp/terraform/helper/acctest"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 	"google.golang.org/api/cloudresourcemanager/v1"
+	resourceManagerV2Beta1 "google.golang.org/api/cloudresourcemanager/v2beta1"
 )
 
 // Test that an IAM binding can be applied to a folder
@@ -245,6 +247,28 @@ func testAccCheckGoogleFolderIamBindingExists(expected *cloudresourcemanager.Bin
 		}
 		return nil
 	}
+}
+
+func getFolderIamPolicyByParentAndDisplayName(parent, displayName string, config *Config) (*cloudresourcemanager.Policy, error) {
+	queryString := fmt.Sprintf("lifecycleState=ACTIVE AND parent=%s AND displayName=%s", parent, displayName)
+	searchRequest := &resourceManagerV2Beta1.SearchFoldersRequest{
+		Query: queryString,
+	}
+	searchResponse, err := config.clientResourceManagerV2Beta1.Folders.Search(searchRequest).Do()
+	if err != nil {
+		if isGoogleApiErrorWithCode(err, 404) {
+			return nil, fmt.Errorf("Folder not found: %s,%s", parent, displayName)
+		}
+
+		return nil, errwrap.Wrapf("Error reading folders: {{err}}", err)
+	}
+
+	folders := searchResponse.Folders
+	if len(folders) != 1 {
+		return nil, fmt.Errorf("expected exactly 1 folder, found %d", len(folders))
+	}
+
+	return getFolderIamPolicyByFolderName(folders[0].Name, config)
 }
 
 func testAccFolderIamBasic(org, fname string) string {
